+++ CAT/data_handling/validate_input.py
@@ -25,8 +25,16 @@ from CAT.data_handling.validation_schemas import (
     core_schema, ligand_schema, qd_schema, database_schema,
     mongodb_schema, bde_schema, qd_opt_schema, crs_schema
 )
+
 from .validate_mol import validate_mol
 from ..utils import validate_path
+from ..attachment.ligand_anchoring import get_functional_groups
+
+try:
+    from dataCAT import Database
+    DATA_CAT = True
+except ImportError:
+    DATA_CAT = False
 
 __all__ = ['validate_input']
 
@@ -73,3 +81,17 @@ def validate_input(s: Settings) -> None:
     # Validate the input cores and ligands
     validate_mol(s.input_cores, 'input_cores', join(path, 'core'))
     validate_mol(s.input_ligands, 'input_ligands', join(path, 'ligand'))
+
+    # Create a dataCAT.Database instance
+    if DATA_CAT:
+        db_path = s.optional.database.dirname
+        s.optional.database.db = Database(path=db_path, **s.optional.database.mongodb)
+    else:
+        s.optional.database.db = False
+
+    # Create RDKit molecules representing functional groups
+    func_groups, split = s.optional.ligand.functional_groups, s.optional.ligand.split
+    if not func_groups:
+        s.optional.ligand.functional_groups = get_functional_groups(None, split)
+    else:
+        s.optional.ligand.functional_groups = get_functional_groups(func_groups)