+++ CAT/mol_utils.py
@@ -38,6 +38,8 @@ from __future__ import annotations
 
 from typing import (Optional, Iterable, Union, Tuple, List)
 
+import numpy as np
+
 from scm.plams import (Molecule, Atom, Bond, MoleculeError, add_to_class)
 from scm.plams.tools.periodic_table import PeriodicTable
 import scm.plams.interfaces.molecule.rdkit as molkit
@@ -166,7 +168,8 @@ def merge_mol(self, mol_list: Union[Molecule, Iterable[Molecule]]) -> None:
 
 @add_to_class(Molecule)
 def separate_mod(self) -> List[Molecule]:
-    """Modified PLAMS function: seperates a molecule instead of a copy of a molecule.
+    """Modified PLAMS function: creates new molecules out of this instance rather than
+    a copy of this instance. Atoms, bonds and properties are *not* copied.
 
     Separate the molecule into connected components.
     Returns is a list of new Molecule instrances (all atoms and bonds are disjoint with
@@ -209,13 +212,30 @@ def separate_mod(self) -> List[Molecule]:
     return frags
 
 
+@add_to_class(Molecule)
+def round_coords(self, decimals: int = 3) -> None:
+    """Round the Cartesian coordinates of this instance to a given precision in decimal digits.
+
+    Performs an inplace update of all atoms in this instance.
+
+    Parameters
+    ----------
+    decimals : int
+        The desired precision in decimal digits.
+
+    """
+    xyz = self.as_array()
+    np.round(xyz, decimals=decimals, out=xyz)
+    self.from_array(xyz)
+
+
 def to_atnum(item: Union[str, int]) -> int:
     """Turn an atomic symbol into an atomic number.
 
     Parameters
     ----------
     item : |int|_ or |str|_
-    An atomic symbol or number.
+        An atomic symbol or number.
 
     Returns
     -------
@@ -233,7 +253,7 @@ def to_atnum(item: Union[str, int]) -> int:
     elif isinstance(item, int):
         return item
 
-    err = "item expects an instance of 'str' or 'int'; observed type: '{}'"
+    err = "the 'item' argument expects an instance of 'str' or 'int'; observed type: '{}'"
     raise TypeError(err.format(item.__class__.__name__))
 
 
@@ -243,7 +263,7 @@ def to_symbol(item: Union[str, int]) -> str:
     Parameters
     ----------
     item : |int|_ or |str|_
-    An atomic symbol or number.
+        An atomic symbol or number.
 
     Returns
     -------
@@ -261,16 +281,16 @@ def to_symbol(item: Union[str, int]) -> str:
     elif isinstance(item, str):
         return item
 
-    err = "item expects an instance of 'str' or 'int'; observed type: '{}'"
+    err = "the 'item' argument expects an instance of 'str' or 'int'; observed type: '{}'"
     raise TypeError(err.format(item.__class__.__name__))
 
 
-def adf_connectivity(plams_mol: Molecule) -> List[str]:
+def adf_connectivity(mol: Molecule) -> List[str]:
     """Create an AMS-compatible connectivity list.
 
     Parameters
     ----------
-    plams_mol : |plams.Molecule|_
+    mol : |plams.Molecule|_
         A PLAMS molecule with :math:`n` bonds.
 
     Returns
@@ -279,16 +299,23 @@ def adf_connectivity(plams_mol: Molecule) -> List[str]:
         An ADF-compatible connectivity list of :math:`n` bonds.
 
     """
+    mol.set_atoms_id()
+
     # Create list of indices of all aromatic bonds
-    rdmol = molkit.to_rdmol(plams_mol)
+    try:
+        rdmol = molkit.to_rdmol(mol)
+    except ValueError:  # Plan B: ignore aromatic bonds
+        bonds = [f'{bond.atom1.id} {bond.atom2.id} {bond.order:.1f}' for bond in mol.bonds]
+        mol.unset_atoms_id()
+        return bonds
+
     aromatic = [bond.GetIsAromatic() for bond in rdmol.GetBonds()]
 
     # Create a list of bond orders; aromatic bonds get a bond order of 1.5
-    plams_mol.set_atoms_id()
-    bond_orders = [(1.5 if ar else bond.order) for ar, bond in zip(aromatic, plams_mol.bonds)]
-    bonds = ['{:d} {:d} {:.1f}'.format(bond.atom1.id, bond.atom2.id, bond.order) for
-             bond, order in zip(plams_mol.bonds, bond_orders)]
-    plams_mol.unset_atoms_id()
+    bond_orders = [(1.5 if ar else bond.order) for ar, bond in zip(aromatic, mol.bonds)]
+    bonds = [f'{bond.atom1.id} {bond.atom2.id} {order:.1f}' for
+             bond, order in zip(mol.bonds, bond_orders)]
+    mol.unset_atoms_id()
 
     return bonds