+++ tests/test_validate_input.py
@@ -5,10 +5,12 @@ from shutil import rmtree
 
 import yaml
 
+from rdkit import Chem
 from scm.plams import (Settings, AMSJob)
 
-from CAT.assertion_functions import assert_eq
+from CAT.assertion_functions import (assert_eq, assert_instance)
 from CAT.data_handling.validate_input import validate_input
+from dataCAT import Database
 
 PATH = 'tests/test_files'
 
@@ -30,19 +32,23 @@ def test_validate_input() -> None:
     ref.database.overwrite = ()
     ref.database.read = ('core', 'ligand', 'qd')
     ref.database.write = ('core', 'ligand', 'qd')
+    ref.database.db = Database(ref.database.dirname, **ref.database.mongodb)
 
     ref.ligand['cosmo-rs'] = False
     ref.ligand.dirname = join(PATH, 'ligand')
-    ref.ligand.functional_groups = None
     ref.ligand.optimize = True
     ref.ligand.split = True
 
     ref.qd.activation_strain = False
     ref.qd.dirname = join(PATH, 'qd')
     ref.qd.dissociate = False
-    ref.qd.optimize = {'job1': AMSJob, 's2': {'description': 'UFF with the default forcefield', 'input': {'uff': {'library': 'uff'}, 'ams': {'system': {'bondorders': {'_1': None}, 'charge': 0}}}}, 's1': {'description': 'UFF with the default forcefield', 'input': {'uff': {'library': 'uff'}, 'ams': {'system': {'bondorders': {'_1': None}, 'charge': 0}}}}, 'job2': AMSJob}  # noqa
+    ref.qd.optimize = {'job1': AMSJob, 's2': {'description': 'UFF with the default forcefield', 'input': {'uff': {'library': 'uff'}, 'ams': {'system': {'bondorders': {'_1': None}}}}}, 's1': {'description': 'UFF with the default forcefield', 'input': {'uff': {'library': 'uff'}, 'ams': {'system': {'bondorders': {'_1': None}}}}}, 'job2': AMSJob}  # noqa
+
+    func_groups = s.optional.ligand.pop('functional_groups')
 
     try:
+        for mol in func_groups:
+            assert_instance(mol, Chem.Mol)
         assert_eq(s.optional, ref)
     finally:
         rmtree(join(PATH, 'ligand'))