+++ CHANGELOG.rst
@@ -5,10 +5,48 @@ Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to `Semantic Versioning <http://semver.org/>`_.
 
+
+0.5.4
+*****
+
+* Minor updates to the logger.
+* Cleaned up CAT.jobs.py.
+* ``check_sys_var()`` is now only called if an ADF-specific Job is requirest.
+* Job hashes are now stored in (and retrieved from) $JN.hash files (plain text).
+* Added a permanent Database_ instance to .optional.database.db.
+* Parsing of functional group SMILES_ strings is now carried out during the Schema_ validation.
+
+.. _Database: https://cat.readthedocs.io/en/latest/7_database.html#class-api
+.. _Schema: https://github.com/keleshev/schema
+
+
+0.5.3
+*****
+
+* Moved Molecule to file exporting (*i.e.* .xyz and .pdb creation) from data-CAT_ to CAT_.
+* Molecules can now be exported to .mol and .mol2 formats (in addition to .pdb and .xyz format).
+* Increased the clarity of many exceptions (see https://github.com/nlesc-nano/CAT/issues/45).
+* Updated the documentation.
+* Introduced a proper logger (see https://github.com/nlesc-nano/CAT/issues/46).
+* Updated data-CAT_ to 0.1.1 (https://github.com/nlesc-nano/data-CAT/pull/12) and
+  nano_CAT_ to 0.1.2 (https://github.com/nlesc-nano/nano-CAT/pull/10).
+
+
+0.5.2
+*****
+
+* Added more tests.
+* Added a more explicit error message to ``_smiles_to_rdmol()``.
+
+
 0.5.1
 *****
 
 * Documentation update.
+* Updated to the ligand dissociation module in nano-CAT_ (see https://github.com/nlesc-nano/nano-CAT/issues/1).
+* Added the ``keep_files`` keyword to the cosmo-rs and ligand dissociation workflows.
+  Default value: ``True``.
+* See https://github.com/nlesc-nano/nano-CAT/pull/9.
 
 
 0.5.0